updated example and decryption based on valid ciphertext
